requirements.txt
asciitree==0.3.3
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
cloudpickle==2.2.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dask==2023.3.2
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fasteners==0.18
Flask==2.2.3
fonttools==4.39.3
fsspec==2023.4.0
importlib-metadata==6.3.0
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
locket==1.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numcodecs==0.11.0
numpy==1.24.2
packaging==23.0
pandas==2.0.0
parso==0.8.3
partd==1.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
six==1.16.0
stack-data==0.6.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
wcwidth==0.2.6
Werkzeug==2.2.3
zarr==2.14.2
zipp==3.15.0
